+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Support editable dependencies from setuptools (experimental).
 * Add `update()` function.
 * Improved input validation.
+* Require Python 3.9+.
 
 ## v0.1.17 (2025-05-13)
 * Respect `JULIAUP_DEPOT_PATH` when searching for Julia using juliaup.